FIxed bugs in array size
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@6861 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ daq/slc/src/node/src/COPPERNode.cc
@@ -26,7 +26,7 @@ const std::string COPPERNode::getSQLFields() const throw()
 {
   std::stringstream ss;
   ss << NSMNode::getSQLFields();
-  for (int i = 0; i < MAX_HSLBS; i++) {
+  for (size_t i = 0; i < MAX_HSLBS; i++) {
     ss << ", hslb_id_" << i << " smallint";
   }
   ss << _sender->getSQLFields();
@@ -37,7 +37,7 @@ const std::string COPPERNode::getSQLLabels() const throw()
 {
   std::stringstream ss;
   ss << NSMNode::getSQLLabels();
-  for (int i = 0; i < MAX_HSLBS; i++) {
+  for (size_t i = 0; i < MAX_HSLBS; i++) {
     ss << ", hslb_id_" << i;
   }
   ss << _sender->getSQLLabels();
@@ -48,7 +48,7 @@ const std::string COPPERNode::getSQLValues() const throw()
 {
   std::stringstream ss;
   ss << NSMNode::getSQLValues();
-  for (int i = 0; i < MAX_HSLBS; i++) {
+  for (size_t i = 0; i < MAX_HSLBS; i++) {
     int id = (_hslb_v[i] != NULL) ? _hslb_v[i]->getID() : -1;
     ss << ", " << id;
   }
