+++ daq/slc/src/node/src/DataReceiverNode.cc
@@ -26,7 +26,7 @@ const std::string DataReceiverNode::getSQLFields() const throw()
   std::stringstream ss;
   ss << NSMNode::getSQLFields()
      << ", script text";
-  for (int i = 0; i < MAX_SENDERS; i++) {
+  for (size_t i = 0; i < MAX_SENDERS; i++) {
     ss << ", sender_id_" << i << " smallint";
   }
   return ss.str();
@@ -37,7 +37,7 @@ const std::string DataReceiverNode::getSQLLabels() const throw()
   std::stringstream ss;
   ss << NSMNode::getSQLLabels()
      << ", script";
-  for (int i = 0; i < MAX_SENDERS; i++) {
+  for (size_t i = 0; i < MAX_SENDERS; i++) {
     ss << ", sender_id_" << i;
   }
   return ss.str();
@@ -48,7 +48,7 @@ const std::string DataReceiverNode::getSQLValues() const throw()
   std::stringstream ss;
   ss << NSMNode::getSQLValues()
      << ", '" << _script << "'";
-  for (int i = 0; i < MAX_SENDERS; i++) {
+  for (size_t i = 0; i < MAX_SENDERS; i++) {
     if (i < _sender_i) {
       ss << "," << _sender_v[i]->getID();
     } else {
